requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.8
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
autopep8==1.5.5
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crcmod==1.7
cryptography==44.0.0
dbus-python==1.3.2
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.9
distro==1.9.0
Django==5.1.5
dnspython==2.7.0
docstring_parser==0.16
executing==2.2.0
filelock==3.16.1
flake8==3.8.4
Flask==3.1.0
Flask-PyMongo==2.3.0
flatbuffers==24.12.23
frozendict==2.4.6
gast==0.6.0
gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz#sha256=1533c75fd582af8102f3c7bb800342521b0964266f723e29ebd5afddc1089c4d
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.76.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.2
google-cloud-language==2.16.0
google-cloud-logging==3.11.3
google-cloud-pubsub==2.29.0
google-cloud-resource-manager==1.14.0
google-cloud-spanner==3.51.0
google-cloud-storage==2.19.0
google-cloud-translate==3.19.0
google-cloud-videointelligence==2.15.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpc-interceptor==0.15.4
grpcio==1.69.0
grpcio-status==1.62.3
h5py==3.12.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.31.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.17.2
jeepney==0.8.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mercurial==6.7.2
ml-dtypes==0.4.1
more-itertools==10.5.0
multitasking==0.0.11
namex==0.0.8
narwhals==1.36.0
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.7.1
peewee==3.17.9
pexpect==4.9.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.6.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==2.2.0
Pygments==2.19.1
PyGObject==3.48.2
PyJWT==2.7.0
pylint==3.3.3
pymongo==4.10.1
pyparsing==3.2.1
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytoolconfig==1.3.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rope==1.13.0
rpds-py==0.22.3
rsa==4.9
SecretStorage==3.3.3
setuptools==75.8.0
shapely==2.0.6
six==1.17.0
snowballstemmer==2.2.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
supervisor==4.2.5
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tink==1.10.0
toml==0.10.2
tomlkit==0.13.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.1
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.1
yapf==0.43.0
yfinance==0.2.56
zipp==3.21.0
